requirements.txt
streamlit
pdf2image
# sudo apt-get update
# sudo apt-get install poppler-utils

langchain
langchain_core
langchain_community
pillow
openai

unstructured-client
unstructured

